TCC Point Confirm API 구현

+++ point/src/main/java/org/example/point/infrastructure/PointRepository.java
@@ -3,7 +3,7 @@ package org.example.point.infrastructure;
 import org.example.point.domain.Point;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PointRepository extends JpaRepository<Point, Integer> {
+public interface PointRepository extends JpaRepository<Point, Long> {
 
     Point findByUserId(Long userId);
 }
